requirements.txt
python-dotenv
python-telegram-bot
SQLAlchemy